feat: 사용자 대화 태그 수정 시 태그 갯수 관련 예외처리 추가(#15)

+++ src/main/java/org/example/tokpik_be/tag/service/TopicTagService.java
@@ -48,6 +48,10 @@ public class TopicTagService {
     public UserTopicTagResponse updateUserTopicTags(long userId, UserTopicTagsRequest request) {
         User user = userQueryService.findById(userId);
 
+        if (request.getTopicTagIds() == null || request.getTopicTagIds().length == 0) {
+            throw new GeneralException(TagException.INVALID_REQUEST);
+        }
+
         userTopicTagRepository.deleteByUserId(user.getId());
 
         for (long tagId : request.getTopicTagIds()) {
